requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
faker==24.4.0
aiohttp==3.9.3
kafka-python==2.0.2